更新gradle及修改获取versionCode代码

+++ app/src/main/java/com/aries/library/fast/demo/entity/UpdateEntity.java
@@ -33,7 +33,7 @@ public class UpdateEntity {
     public float saturation = 1.0f;
 
     public boolean isSuccess() {
-        int code = FastUtil.getVersionCode(App.getContext());
+        long code = FastUtil.getVersionCode(App.getContext());
         String name = FastUtil.getVersionName(App.getContext());
         if (versionCode > code || (versionCode == code && name.compareTo(versionName) < 0)) {
             return true;
